+++ code/src/drivers/PostgreSQL/PostgreSQLConnection.cpp
@@ -1090,7 +1090,7 @@ Strings PostgreSQLConnection::extractStatements(const Strings& sqlBatch) const
         if (functionHeader && !functionBody && matchFunctionBodyStart.matches(row)) {
             functionBody = true;
             functionHeader = false;
-            delimiter = matches[0].value;
+            delimiter = matches[(size_t)0].value;
             statement << row << "\n";
             continue;
         }
